+++ server/src/infrastructure/repositories/competition-sqlite.ts
