+++ server/src/infrastructure/repositories/manual-record-sqlite.ts
