+++ server/src/infrastructure/repositories/participant-sqlite.ts
