+++ server/src/infrastructure/repositories/record-sqlite.ts
